### 2025-08-27 #260 @ldbo
Phileas — reproducible automation for hardware security experiments
Submitting Author: @ldbo
All current maintainers: @ldbo
Package Name: Phileas
One-Line Description of Package: Phileas — reproducible automation for hardware security experiments
Repository Link: https://github.com/ldbo/phileas
Version submitted: v0.5
EiC: TBD
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

Phileas is a Python package that improves the sustainability of hardware security experiments. This domain heavily relies on automated experiments, which rely on laboratory instruments to gather many measurements on the device under test. Phileas has different features that improve the scientific sustainability — including reproducibility, maintainability, reliability and transparency — of this process.

 - It uses simple, human-readable configuration files to describe complex experiments. This improves the readability and maintainability of the research process, while facilitating information exchange between researchers.
 - It provides composable parameter generation strategies that ease the design of simple as well as complex experiments.
 - It provides an abstraction layer for configuring heterogeneous instruments and retrieving their state.

Combined, these features enable automation of data-acquisition experiments, producing more reliable procedures that are easy to replicate, extend, and share.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

It targets experimental researchers in the field of hardware cybersecurity, or hardware designers willing to evaluate the security of their circuits. They can use it when acquiring datasets, usually to carry out side-channel or fault injection attacks. It has been built in a generic manner, which could make it relevant to other fields requiring the automated configuration of test and measurement instruments.

Some tools focus on instruments and drivers ([ChipWhisperer](https://github.com/newaetech/chipwhisperer), [ChipSHOUTER](https://github.com/newaetech/ChipSHOUTER), [Scaffold](https://github.com/Ledger-Donjon/scaffold), [pypdm](https://github.com/Ledger-Donjon/pypdm)). Others focus on data processing ([lascar](https://github.com/Ledger-Donjon/lascar), [SCALib](https://github.com/simple-crypto/SCALib), [SCAred](https://github.com/eshard/scared), [secbench](https://github.com/CEA-Leti/secbench)) or provide GUIs for specific experiments ([laserstudio](https://github.com/Ledger-Donjon/laserstudio)). However, to my knowledge, there is no tool that automates complex data acquisition procedures.

https://github.com/pyOpenSci/software-submission/issues/256, which was reviewed by @eliotwrobson.

- [x] The package is deposited in a long-term repository with the DOI: 10.5281/zenodo.16943365
