### 2025-08-27 #260 @ldbo
Two notes about the JOSS submission:
- A Zenodo DOI is reserved, however I have not published the package yet. JOSS states that the repository should be archived after the review process is completed. Can I wait for the review process to be done?
- The JOSS paper is in the `/joss-paper/` directory of the [`joss-paper`](https://github.com/ldbo/phileas/tree/joss-paper) branch. You can find the PDF in the artifacts of [this workflow](https://github.com/ldbo/phileas/actions/workflows/joss-paper.yaml). `joss-paper` will be a short-lived branch never merged with `master`. It follows the JOSS submission requirements, but is it ok for pyOpenSci?
