### 2025-08-29 #260 @eliotwrobson
@ldbo those are both fine! We only really need to worry about the parts of the JOSS process once the review is completed. I'll start looking around for an editor.
