### 2026-02-02 #260 @crhea93
Hi @ldbo! I'll be taking over as editor in chief for this review. Below I'll be posting the pre check review list where I'll do a first run through of the package :) While you are responding to any potential comments, I'll be starting to look for an editor.
